requirements.txt
appdirs==1.4.0
Django==1.10.5
django-allauth==0.30.0
django-appconf==1.0.1
django-crispy-forms==1.6.1
django-image-cropping==1.0.4
djangorestframework==3.5.4
easy-thumbnails==2.3
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
python-openid==2.2.5
requests==2.13.0
requests-oauthlib==0.7.0
six==1.10.0
